Added :root pseudo class selector and rem units.
This illustrates the use of rem units. 1 rem is equivalent to the current font-size
of the root element(<html>). Hence becauase a rem is relative to the root element, it is ideal
for setting the font-sizes of other elements using rem, then set other properties (i.e. padding,
margin, border-radius, etc.) using ems, which will be relative to the local font-size of their element,
which, would have been set by rems.